change ChromosomeView.ts
